requirements.txt
pyaarlo
requests
python-dotenv
google-generativeai
